qmlsink: defer context initilization until beforeSynchronizing()
If a qmlsink is dynamically created after a QML application has already
started, the scene graph may have already been constructed, so the
onSceneGraphInitialized() callback is not initialized. When the renderer
attempts to update the node, it fails on a missing context.

Do not register for onSceneGraphInitialized, but for
beforeSynchronizing instead, which will get called just before scene
graph synchronization. The GL context should be valid at this point.

Signed-off-by: Perry Hung <iperry@gmail.com>